# apps/api/gpt/openapi.yaml
openapi: 3.1.0
info:
  title: TimTracker Query API
  description: |
    API for querying TimTracker tables and summarizing health data. 
    Exposes both raw table access and analytic endpoints. 
    Use /api/schema to discover tables, columns, and endpoints. 
    For most use cases, prefer analytic endpoints (e.g., /api/sleep) over raw tables.
  version: '2.0'
servers:
  - url: https://timtracker-api.vercel.app
paths:
...
  /api/schema/{table}/{column}/values:
    get:
      operationId: getColumnValues
      summary: Get unique values for a column
      parameters:
        - name: table
          in: path
          required: true
          schema:
            type: string
        - name: column
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Unique values for the column
          content:
            application/json:
              schema:
                type: object
                properties:
                  values:
                    type: array
                    items:
                      type: string
  /api/query:
    post:
      operationId: postFlexibleQuery
      summary: Flexible query for any table
      description: |
        Query any table with custom columns, filters, order, and limit. 
        For raw tables, see /api/schema for recommended analytic endpoints.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                table:
                  type: string
                columns:
                  type: array
                  items:
                    type: string
                filters:
                  type: array
                  items:
                    type: object
                    properties:
                      column:
                        type: string
                      op:
                        type: string
                        enum: ['=', '>', '<', '>=', '<=', '!=', 'LIKE']
                      value:
                        type: string
                order_by:
                  type: string
                limit:
                  type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of rows matching the query
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
...
  /api/people:
    get:
      summary: List people
      operationId: listPeople
      parameters:
        - in: query
          name: q
          required: false
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of people
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object

    post:
      summary: Create or update a person
      operationId: createPerson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                note:
                  type: string
                gender:
                  type: string
                  enum: [male, female, other]
                importance:
                  type: integer
                alive:
                  type: boolean
                tag:
                  type: array
                  items:
                    type: string
                relationships:
                  type: array
                  items:
                    type: integer
              required:
                - name
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Person ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
  /api/people/{id}:
    patch:
      summary: Update a person
      operationId: updatePerson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                note:
                  type: string
                gender:
                  type: string
                  enum: [male, female, other]
                importance:
                  type: integer
                alive:
                  type: boolean
                tag:
                  type: array
                  items:
                    type: string
                relationships:
                  type: array
                  items:
                    type: integer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Person updated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token required to authenticate requests. 
        Use the GPT API key configured in the ChatGPT connector settings.
...